// packages/twenty-server/src/modules/calendar/calendar-event-import-manager/types/calendar-event-save-operations.ts
import { type QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';

import { type QueryDeepPartialEntityWithNestedRelationFields } from 'src/engine/twenty-orm/entity-manager/types/query-deep-partial-entity-with-nested-relation-fields.type';

import { type CalendarChannelEventAssociationWorkspaceEntity } from 'src/modules/calendar/common/standard-objects/calendar-channel-event-association.workspace-entity';
import { type CalendarEventWorkspaceEntity } from 'src/modules/calendar/common/standard-objects/calendar-event.workspace-entity';
import { type FetchedCalendarEventParticipant } from 'src/modules/calendar/common/types/fetched-calendar-event';

export type CalendarEventSaveOperations = {
  calendarEventsToInsert: QueryDeepPartialEntityWithNestedRelationFields<CalendarEventWorkspaceEntity>[];
  calendarEventsToUpdate: {
    criteria: string;
    partialEntity: QueryDeepPartialEntity<CalendarEventWorkspaceEntity>;
  }[];
  associationsToInsert: Pick<
    CalendarChannelEventAssociationWorkspaceEntity,
    | 'calendarEventId'
    | 'eventExternalId'
    | 'calendarChannelId'
    | 'recurringEventExternalId'
  >[];
  associationsToUpdate: {
    criteria: string;
    partialEntity: QueryDeepPartialEntity<CalendarChannelEventAssociationWorkspaceEntity>;
  }[];
  participantsToCreate: (FetchedCalendarEventParticipant & {
    calendarEventId: string;
  })[];
  participantsToUpdate: (FetchedCalendarEventParticipant & {
    calendarEventId: string;
  })[];
};